consumers: removed hk2 factories
