+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/consumer/sender/http/Http2ClientHolder.java
@@ -8,7 +8,7 @@ public class Http2ClientHolder {
 
     private final HttpClient http2Client;
 
-    Http2ClientHolder(HttpClient http2Client) {
+    public Http2ClientHolder(HttpClient http2Client) {
         this.http2Client = http2Client;
     }
