+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/di/config/MessageConfiguration.java
@@ -1,0 +1,17 @@
+package pl.allegro.tech.hermes.consumers.di.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import pl.allegro.tech.hermes.common.config.ConfigFactory;
+import pl.allegro.tech.hermes.common.message.undelivered.UndeliveredMessageLog;
+import pl.allegro.tech.hermes.consumers.message.undelivered.UndeliveredMessageLogPersister;
+
+@Configuration
+public class MessageConfiguration {
+
+    @Bean
+    public UndeliveredMessageLogPersister undeliveredMessageLogPersister(UndeliveredMessageLog undeliveredMessageLog,
+                                                                         ConfigFactory configFactory) {
+        return new UndeliveredMessageLogPersister(undeliveredMessageLog, configFactory);
+    }
+}
