+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/supervisor/workload/ConsumerWorkloadRegistryType.java
@@ -3,7 +3,7 @@ package pl.allegro.tech.hermes.consumers.supervisor.workload;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
-enum ConsumerWorkloadRegistryType {
+public enum ConsumerWorkloadRegistryType {
 
     HIERARCHICAL("hierarchical"),
     FLAT_BINARY("flat-binary");
